refactor: skimage/astropy/matplotlib API drift
Phase 4: get imports working against numpy 2.x / scipy 1.16 / scikit-image
0.22 / matplotlib >=3.9 / astropy >=6 / pyregion 2.2 / Python 3.11.

* numpy: replace bare np.int / np.float / np.complex aliases (removed in
  1.24) with the Python builtins they originally aliased.
* scikit-image: skimage.morphology.watershed -> skimage.segmentation.watershed.
* matplotlib: qt4agg backend rename, inline replacement for the removed
  matplotlib.mlab.dist_point_to_segment, inline AnchoredEllipse class
  (removed in matplotlib 3.8).
* scipy: rewrite leftover 'from scipy.ndimage import measurements'
  imports (Phase 2's verify regex didn't catch this form) onto the
  ndimage namespace.
* pkg_resources -> importlib.resources (setuptools 81+ no longer ships
  pkg_resources by default, so this turned into a blocker).

`python -c "import libwise, wise; print(libwise.get_version(), wise.get_version())"` succeeds.

See MIGRATION_NOTES.md for per-substitution rationale.

+++ packages/libwise/src/libwise/nputils.py
@@ -1092,7 +1092,7 @@ def crop_threshold(array, threashold=0, crop_mask=None, output_index=False):
 
 def nextpow2(n):
     '''get the next power of 2 that's greater than n'''
-    return 2 ** np.ceil(np.log2(n)).astype(np.int)
+    return 2 ** np.ceil(np.log2(n)).astype(int)
 
 
 def angle(v1, v2):
@@ -1862,7 +1862,7 @@ def movingaverage(interval, window_size, boundary='zero', mode='valid'):
 def gaussian_support(sigma=None, width=None, nsigma=4):
     if sigma is None:
         sigma = gaussian_fwhm_to_sigma(width)
-    return np.int(np.ceil(2 * nsigma * sigma))
+    return int(np.ceil(2 * nsigma * sigma))
 
 
 def gaussian_sigma_to_fwhm(sigma):
@@ -1928,7 +1928,7 @@ def make_callable(obj):
 
 def _get_next_oddeven(n, testeven):
     if isinstance(n, np.ndarray):
-        if not n.dtype == np.int:
+        if not n.dtype == int:
             n = n.astype(int)
         if (n % 2 == testeven).any():
             m = n.copy()
@@ -2056,7 +2056,7 @@ def fitgaussian(data, params, base_null=True):
 # def modelfit(data):
 
 #     def errorfunction(*p):
-#         indices = np.indices(np.array(data.shape, dtype=np.int))
+#         indices = np.indices(np.array(data.shape, dtype=int))
 #         p = [0, p[2], p[0], p[1], p[3], p[3], 0]
 
 #         err = np.ravel(data - gaussian_fct(*p)(indices))
